backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
mysqlclient==2.2.0
Pillow>=10.4.0
python-decouple==3.8
daphne==4.0.0
